pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
